2 part of main screen and adaptive header
